Reorganize Binance common client modules
