+++ nautilus_trader/adapters/binance/data.py
